予約者情報をコンポーネント化(#136)
