+++ src/components/reservations/dialog/address/script.ts
@@ -3,7 +3,7 @@ import { ReservationListSeat } from "@/entity/reservation-list-seat";
 
 export default Vue.extend({
   props: {
-    seat: {
+    reserver: {
       required: true,
       type: Object as PropType<ReservationListSeat>
     }
