+++ src/components/reservations/dialog/address/template.html
@@ -8,15 +8,15 @@
     <section class="modal-card-body">
         <div class="item">
             <h4><i class="fas fa-user"></i>お名前</h4>
-            <p>{{ seat.reserver_name }}</p>
+            <p>{{ reserver.reserver_name }}</p>
         </div>
         <div class="item">
             <h4><i class="fas fa-envelope"></i>メールアドレス</h4>
-            <p><a class="button is-info is-inverted" href="mailto:gamoyon@email.com">{{ seat.mail }}</a></p>
+            <p><a class="button is-info is-inverted" href="mailto:gamoyon@email.com">{{ reserver.mail }}</a></p>
         </div>
         <div class="item">
             <h4><i class="fas fa-phone"></i>電話番号</h4>
-            <p>{{ seat.tel }}</p>
+            <p>{{ reserver.tel }}</p>
         </div>
     </section>
