+++ src/components/reservations/list-item/style.scss
@@ -4,13 +4,3 @@
     margin-top: 3px;
     margin-bottom: 3px;
 }
-.card {
-    margin-bottom: 18px;
-}
-.reserver {
-    font-weight: 700;
-    font-size: 1.6rem;
-}
-.fas {
-    margin-right: 4px;
-}
