+++ src/components/reserver/card/ReserverCard.vue
@@ -1,0 +1,3 @@
+<template src="./template.html"></template>
+<script src="./script.ts" lang="ts"></script>
+<style src="./style.scss" lang="scss" scoped></style>
