+++ src/components/reserver/card/style.scss
@@ -1,0 +1,11 @@
+.card {
+    margin-bottom: 18px;
+
+    .reserver {
+        font-weight: 700;
+        font-size: 1.6rem;
+    }
+    .fas {
+        margin-right: 4px;
+    }
+}
