+++ src/components/reserver/card/template.html
@@ -1,0 +1,17 @@
+<div class="card">
+    <div class="card-content">
+        <h3 class="reserver">
+            {{ reserver.reserver_name | formatReserver(reserver.number_of_reservations) }}
+        </h3>
+        <p>座席番号: {{ reserver.seat_no }}</p>
+        <p>{{ reserver.comment }}</p>
+        <div class="buttons is-right">
+            <button type="button" class="button is-info is-inverted" @click="handleShowAddressDialog()">
+                <i class="fas fa-address-card"></i>連絡先を表示
+            </button>
+            <button type="button" class="button is-danger is-inverted" @click="handleCancel()">
+                <i class="fas fa-times"></i>予約を取消
+            </button>
+        </div>
+    </div>
+</div>
